agreementversion controller and service class refactor
Renamed controller
moved service method to existing class
stremlining with existing method
