+++ prime-dotnet-webapi/Services/AgreementService.cs
@@ -117,7 +117,7 @@ namespace Prime.Services
             var agreement = new Agreement
             {
                 EnrolleeId = enrolleeId,
-                AgreementVersionId = await FetchNewestAgreementVersionIdOfType(dto.NewestAssignedAgreement.Value),
+                AgreementVersionId = (await FetchNewestAgreementVersionOfTypeAsync(dto.NewestAssignedAgreement.Value)).Id,
                 LimitsConditionsClause = LimitsConditionsClause.FromAgreementNote(dto.AccessAgreementNote),
                 CreatedDate = DateTimeOffset.Now
             };
@@ -278,6 +278,18 @@ namespace Prime.Services
             return signedAgreement;
         }
 
+        public async Task<IEnumerable<AgreementVersionViewModel>> GetLatestEnrolleeAgreementVersionsAsync()
+        {
+            var agreementVersionList = new List<AgreementVersion>();
+            foreach (var type in AgreementTypeExtensions.EnrolleeAgreementTypes())
+            {
+                agreementVersionList.Add(await FetchNewestAgreementVersionOfTypeAsync(type));
+            }
+            return agreementVersionList.AsQueryable()
+            .ProjectTo<AgreementVersionViewModel>(_mapper.ConfigurationProvider)
+            .ToList();
+        }
+
         /// <summary>
         /// Renders the HTML text of the Agreement for viewing on the frontend.
         /// </summary>
@@ -292,22 +304,21 @@ namespace Prime.Services
             }
         }
 
-        private async Task<int> FetchNewestAgreementVersionIdOfType(AgreementType type)
-        {
-            return await _context.AgreementVersions
-                .AsNoTracking()
-                .OrderByDescending(a => a.EffectiveDate)
-                .Where(a => a.AgreementType == type)
-                .Select(a => a.Id)
-                .FirstAsync();
-        }
-
         private async Task<string> GetOrganizationName(int organizationId)
         {
             return await _context.Organizations
                 .Where(o => o.Id == organizationId)
                 .Select(o => o.Name)
                 .SingleAsync();
+        }
+
+        private async Task<AgreementVersion> FetchNewestAgreementVersionOfTypeAsync(AgreementType type)
+        {
+            return await _context.AgreementVersions
+            .AsNoTracking()
+            .Where(av => av.AgreementType == type)
+            .OrderByDescending(av => av.EffectiveDate)
+            .FirstOrDefaultAsync();
         }
     }
 }
