+++ prime-dotnet-webapi/Services/interfaces/IAgreementService.cs
@@ -32,5 +32,7 @@ namespace Prime.Services
         Task<SignedAgreementDocument> GetSignedAgreementDocumentAsync(int agreementId);
 
         Task<SignedAgreementDocument> AddSignedAgreementDocumentAsync(int agreementId, Guid documentGuid);
+
+        Task<IEnumerable<AgreementVersionViewModel>> GetLatestEnrolleeAgreementVersionsAsync();
     }
 }
