--- prime-dotnet-webapi/Services/interfaces/IAgreementVersionService.cs
@@ -1,11 +1,0 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
-using Prime.ViewModels;
-
-namespace Prime.Services
-{
-    public interface IAgreementVersionService
-    {
-        Task<IEnumerable<AgreementVersionViewModel>> GetLatestEnrolleeAgreementVersionsAsync();
-    }
-}
